Add Cargo.lock file into source control
The official Rust book 'The Cargo Book' states the following:

The purpose of a Cargo.lock lockfile is to describe the state of the
world at the time of a successful build. Cargo uses the lockfile to
provide deterministic builds on different times and different systems,
by ensuring that the exact same dependencies and versions are used as
when the Cargo.lock file was originally generated.

This property is most desirable from applications and packages which are
at the very end of the dependency chain (binaries). As a result, it is
recommended that all binaries check in their Cargo.lock.