Refactor loop logic in construct_templates()
Assume (and please validate this) that a window is defined as starting from
index 0 of ts_data and only iterating until there are no more "full" windows (i.e.
a window may never have fewer than window_size elements).

The following would demonstrate each iteration of a loop needed to iterate
over each unique window of the data:

window_size = 2
ts_data.len() = 5

12345 <- ts_data
12    <- window_one   i=0
 23   <- window_two   i=1
  34  <- window_three i=2
   45 <- window_four  i=3

So this means we need the loop to iterate 4 times to capture each unique window
of the data, or 0 to 3. Rust's range syntax enforces the formula start ≤ x < end.
So we can write 0 to 3 in Rust's range syntax as 0..4 .

Lets try it with:
window_size = 1
ts_data.len() = 3

123 <- ts_data
1   <- window_one   i=0
 2  <- window_two   i=1
  3 <- window_three i=2

This time the loop needed to iterate three times, or 0 to 2. Again, with Rust's
range syntax this is represented as 0..3

In the original code this logic is expressed as the function window_size..ts_data.len() + 1

If we apply the original formula for calculating how many iterations we need we get:
  Example 1: 2..6
  Example 2: 1..4

The inner loop then applies logic to revert the starting index back to zero so by this
point it should be clear that we can clean up the looping logic to something a bit clearer
such as the following:

let num_windows = ts_data.len() - window_size + 1;
return (0..num_windows)
    .map(|x| ts_data[x..x + window_size].to_vec())
    .collect::<Vec<Vec<f32>>>();