Improve clarity of template matching logic
The goal of the get_matches() function is to return (2 * the number of
pairs of template vectors whos chebyshev distance is less than r). By
rewriting this using iterators and the .combinations() method from the
itertools crate, we can remove (at least visually) the double-nested
for-loop code in favor of a clearer functional-style loop.