Remove needless vector creation
By calling .collect() on the iterator over 'data', a new Vec<f32> must
be created and populated with data. By omitting the call to collect and
instead using the .fold() method on the iterator, we avoid any
allocation or creation of a secondary vector entirely, and simply
allow the .fold() method to keep track of the sum as the closure is
applied over the elements of the original vector.