Pass vector as slice instead of by value
In rust, a function that takes a 'Vec<f32>' parameter is passing the
entire vector by value, which may require allocating memory and copying
the elements if the vector needs to be moved. This can have an impact on
performance, especially for large vectors, as it involves additional
memory allocations and copying.

Passing an &[f32] as a function argument has minimal performance
overhead. Slices are lightweight and simply provide a reference to the
original data. They avoid the need for memory allocation or copying,
making them efficient in terms of memory and time.